+++ src/CLI/Command.hs
@@ -48,7 +48,6 @@ import System.Exit (exitFailure)
 import System.FilePath ((</>))
 import Types (Database)
 import qualified Types
-import UnitConversion (defaultUnitConfig)
 
 -- | Default output format for different command types
 defaultFormat :: Command -> OutputFormat
@@ -229,7 +228,7 @@ executeDbCommand fmt _globalOpts database = \case
 -- | Execute activity info command
 executeActivityCommand :: OutputFormat -> Database -> T.Text -> IO ()
 executeActivityCommand fmt database uuid =
-    case Service.getActivityInfo defaultUnitConfig database uuid of
+    case Service.getActivityInfo database uuid of
         Left err -> reportServiceError err
         Right result -> outputResult fmt result
 
@@ -744,7 +743,8 @@ executeFlowMappingCommand fmt database manager opts = do
                     reportError $ "Method not found: " ++ T.unpack (mappingMethodId opts)
                     exitFailure
                 (method : _) -> do
-                    mappings <- mapMethodToFlows database method
+                    cmap <- DM.getMergedCompartmentMap manager
+                    mappings <- mapMethodToFlows cmap database method
                     let stats = computeMappingStats mappings
                         totalMatched = msTotal stats - msUnmatched stats
                         coverage =
@@ -770,9 +770,6 @@ executeFlowMappingCommand fmt database manager opts = do
                             putStrLn $ "  by CAS:     " ++ show (msByCAS stats)
                             putStrLn $ "  by Name:    " ++ show (msByName stats)
                             putStrLn $ "  by Synonym: " ++ show (msBySynonym stats)
-                            when (msByFuzzy stats > 0) $
-                                putStrLn $
-                                    "  by Fuzzy:   " ++ show (msByFuzzy stats)
                             when (msByProxy stats > 0) $
                                 putStrLn $
                                     "  by Proxy:   " ++ show (msByProxy stats)
@@ -829,7 +826,6 @@ executeFlowMappingCommand fmt database manager opts = do
                                         , "byCAS" .= msByCAS stats
                                         , "byName" .= msByName stats
                                         , "bySynonym" .= msBySynonym stats
-                                        , "byFuzzy" .= msByFuzzy stats
                                         , "byProxy" .= msByProxy stats
                                         , "unmatched" .= msUnmatched stats
                                         , "dbBiosphereFlows" .= dbBioCount
@@ -878,7 +874,6 @@ strategyText ByUUID = "uuid"
 strategyText ByCAS = "cas"
 strategyText ByName = "name"
 strategyText BySynonym = "synonym"
-strategyText ByFuzzy = "fuzzy"
 strategyText ByProxy = "proxy"
 strategyText NoMatch = "none"